web/requirements.txt
Flask==2.0.1
python-dotenv
Werkzeug
kafka-python
requests==2.26.0
